REPLSystem: step: change deltatime type to float

+++ include/lua/REPLSystem.h
@@ -21,7 +21,7 @@ class REPLSystem : public System
 public:
 	REPLSystem(EntityManager *em, ComponentManager *cm);
 	virtual ~REPLSystem();
-	void step(unsigned int dt, EntityManager* em);
+	void step(float dt, EntityManager* em);
 
 private:
 	/* Check if stdin is readable, and if so, read it and return the input */
